Clean up ParseEventsF5V1 switch statement based on sample data.
F5V1 introduces a new Hypopnea variant (event 8) with 2 data bytes.
This was previously seen in F5V3, and it's unclear how this is
different from the usual Hypopnea event (event 7, 1 data byte).
Parsing used to stop when it encountered this event, but now it
can continue through the end of the session.

F5V1 also adds a leak value in its periodic stats, also previously
seen in F5V3. PB start time and duration are now correct.

As with F5V0, now VS, LL, and EPAP adjustment are correctly parsed,
and timestamps and durations are now accurate.

The handlers were essentially copied from F5V3 and F5V0 for
consistency and cross-checking.