Shift PRS1 statistics left 2 minutes on import.
PRS1 statistics are reported at the end of a 2-minute interval,
but OSCAR assumes the timestamp for the statistics is their
starting timestamp.

This resolves the issue where statistics wouldn't appear for the
first two minutes of a session, and then would consistently be
2 minutes delayed.

There are still some edge cases to deal with, but the general
case is much improved.