Allow limited reordering of serial port replay.
Now a replayed read() will return the response that follows the
matching write().

When calls are made in the same order as they were during recording,
this will have no effect, and the original ordering will be replayed.

However, minor changes to the code should still result in sensible
replay until a new recording can be made.