Fix OSCAR crash when session not found in Journal file
While preparing bookmark HTML for right sidebar, OSCAR was trying to obtain
the first MT_JOURNAL session, but was finding a null pointer and then not checking
that the pointer was null. However, I don't know under what conditions this
situation could happen in the database.  At least now it won't crash OSCAR.