Add support for yet another Viatom/Wellue filename variation.
Apparently the Android app is now exporting files with timestamps of the form
"YYYY-MM-DD hh:mm:ss". It turns out that ":" is not a valid character on
macOS, so Mac users using version 2.72 of the Android app will need to rename
their files to end with "YYYYMMDDhhmmss" in order to select and import them.
Windows and Linux won't.

Fortunately the intersection of Android users and Mac users is relatively small.
And this may be reverted in a future version of the Android app.

Also clean up some competing release notes edits.