Remove noisy debug message in PRS1 loader.
In theory it was useful to know why files were skipped, but in practice we
normally skip all but the most recent files, often thousands. If you had the
misfortune to have the debug pane open, you'd have to wait for an extra
minute or two for all those messages to scroll by.

If and when we can limit import scanning to files that are new or changed,
then the debug message will be useful again.